Type missmatch

+++ src/MagicMapReferenceTester/PersonModel.cs
@@ -13,6 +13,6 @@ public class PersonModel
 {
    public string Name { get; set; } = null!;
 
-   public int Age { get; set; }
+   public double Age { get; set; }
    
 }
